added ability to delete books
