+++ package.json
@@ -24,25 +24,25 @@
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
-    "@hookform/resolvers": "^3.9.1",
+    "@hookform/resolvers": "^3.10.0",
     "@neondatabase/serverless": "^0.10.4",
     "@radix-ui/react-avatar": "^1.1.2",
-    "@radix-ui/react-checkbox": "^1.1.2",
+    "@radix-ui/react-checkbox": "^1.1.3",
     "@radix-ui/react-context-menu": "^2.2.5",
-    "@radix-ui/react-dialog": "^1.1.2",
-    "@radix-ui/react-dropdown-menu": "^2.1.4",
-    "@radix-ui/react-label": "^2.1.0",
+    "@radix-ui/react-dialog": "^1.1.5",
+    "@radix-ui/react-dropdown-menu": "^2.1.5",
+    "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-scroll-area": "^1.2.2",
-    "@radix-ui/react-select": "^2.1.2",
+    "@radix-ui/react-select": "^2.1.5",
     "@radix-ui/react-separator": "^1.1.1",
-    "@radix-ui/react-slot": "^1.1.0",
+    "@radix-ui/react-slot": "^1.1.1",
     "@radix-ui/react-switch": "^1.1.2",
     "@radix-ui/react-tabs": "^1.1.2",
-    "@radix-ui/react-toast": "^1.2.4",
+    "@radix-ui/react-toast": "^1.2.5",
     "@radix-ui/react-toggle": "^1.1.1",
-    "@radix-ui/react-tooltip": "^1.1.6",
+    "@radix-ui/react-tooltip": "^1.1.7",
     "@t3-oss/env-nextjs": "^0.10.1",
-    "@tanstack/react-query": "^5.62.11",
+    "@tanstack/react-query": "^5.65.1",
     "@trpc/client": "11.0.0-rc.682",
     "@trpc/next": "11.0.0-rc.682",
     "@trpc/react-query": "11.0.0-rc.682",
@@ -51,42 +51,42 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.0.0",
-    "dotenv": "^16.4.5",
-    "drizzle-orm": "^0.38.3",
+    "dotenv": "^16.4.7",
+    "drizzle-orm": "^0.38.4",
     "fuse.js": "^7.0.0",
-    "geist": "^1.3.0",
+    "geist": "^1.3.1",
     "lucide-react": "^0.468.0",
     "next": "15.1.3",
     "next-auth": "5.0.0-beta.25",
-    "nuqs": "^2.3.0",
-    "postgres": "^3.4.4",
+    "nuqs": "^2.3.1",
+    "postgres": "^3.4.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-hook-form": "^7.53.2",
+    "react-hook-form": "^7.54.2",
     "react-resizable-panels": "^2.1.7",
     "server-only": "^0.0.1",
     "superjson": "^2.2.2",
-    "tailwind-merge": "^2.5.5",
+    "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.23.8"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
-    "@tanstack/react-query-devtools": "^5.62.11",
-    "@types/eslint": "^8.56.10",
-    "@types/node": "^20.14.10",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^8.1.0",
-    "@typescript-eslint/parser": "^8.1.0",
-    "drizzle-kit": "^0.30.1",
-    "eslint": "^8.57.0",
-    "eslint-config-next": "^15.0.1",
+    "@tanstack/react-query-devtools": "^5.65.1",
+    "@types/eslint": "^8.56.12",
+    "@types/node": "^20.17.16",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "^18.3.5",
+    "@typescript-eslint/eslint-plugin": "^8.22.0",
+    "@typescript-eslint/parser": "^8.22.0",
+    "drizzle-kit": "^0.30.3",
+    "eslint": "^8.57.1",
+    "eslint-config-next": "^15.1.6",
     "eslint-plugin-drizzle": "^0.2.3",
-    "postcss": "^8.4.39",
-    "prettier": "^3.3.2",
-    "prettier-plugin-tailwindcss": "^0.6.5",
-    "tailwindcss": "^3.4.3",
-    "typescript": "^5.5.3"
+    "postcss": "^8.5.1",
+    "prettier": "^3.4.2",
+    "prettier-plugin-tailwindcss": "^0.6.11",
+    "tailwindcss": "^3.4.17",
+    "typescript": "^5.7.3"
   },
   "ct3aMetadata": {
     "initVersion": "7.38.1"
