+++ src/app/library/_components/BookList.tsx
@@ -1,15 +1,14 @@
 "use client";
-import { parseAsInteger, useQueryState } from "nuqs";
+import { parseAsInteger } from "nuqs";
 import { api } from "~/trpc/react";
-import ItemList from "./ItemList";
 import { useQueryLocalStorage } from "~/app/_components/hooks/useLocalStorage";
 
-import { getPagesLeft } from "~/utils";
+import { ScrollArea } from "~/components/ui/scroll-area";
+import BookItem from "./BookItem";
 
 export default function CharacterList() {
   const [characters] = api.character.getMyCharacters.useSuspenseQuery();
   const [characterId] = useQueryLocalStorage("character", parseAsInteger);
-  const [bookId, setBookId] = useQueryState("book", parseAsInteger);
 
   if (!characterId) return <div>Select a character to view their books</div>;
 
@@ -18,12 +17,15 @@ export default function CharacterList() {
   const characterName = characters.find((c) => c.id === characterId)?.name;
 
   return (
-    <ItemList
-      books={books ?? []}
-      onClick={(b) => setBookId(() => b.id)}
-      selectedItemId={bookId}
-      title={`${characterName}'s Books`}
-      info={(b) => getPagesLeft(b)}
-    />
+    <ScrollArea>
+      <div className="p-4">
+        <h4 className="mb-4 flex justify-between text-sm font-medium leading-none">
+          <span>{`${characterName}'s Books`}</span> <span>pages left</span>
+        </h4>
+        {(books ?? []).map((book) => (
+          <BookItem key={book.id} book={book} />
+        ))}
+      </div>
+    </ScrollArea>
   );
 }
